+++ CHANGELOG.md
@@ -11,6 +11,10 @@ This project adheres to the [Semantic Versioning 2.0 Specification](https://semv
 - JSON, JSONC, and JSON5 are no longer supported as data file formats.
 - Supplying a custom template renderer is no longer supported.
 
+### ✨ New
+
+- `puppeteerLaunchOptions` can now be set in the config file.
+
 ## 0.2.0 - February 01, 2023
 
 ### ⚠️ Breaking